Refactor IT with fluent assertion
